+++ src/pages/home/index.tsx
@@ -1,7 +1,20 @@
+import Input from '@components/input'
+
 export default function HomePage() {
   return (
-    <div>
-      <p>HELLO WORLD!</p>
-    </div>
+    <>
+      <Input inputType="name" placeholder="name" inputAccessedFor="login" />
+      <Input inputType="email" placeholder="email" inputAccessedFor="login" />
+      <Input
+        inputType="password"
+        placeholder="password"
+        inputAccessedFor="signup"
+      />
+      <Input
+        inputType="password_check"
+        placeholder="password_check"
+        inputAccessedFor="signup"
+      />
+    </>
   )
 }
